+++ src/test/java/org/ojalgo/optimisation/integer/OptimisationIntegerTests.java
@@ -21,14 +21,12 @@
  */
 package org.ojalgo.optimisation.integer;
 
-import org.ojalgo.concurrent.Parallelism;
 import org.ojalgo.optimisation.integer.IntegerStrategy.ConfigurableStrategy;
 
 abstract class OptimisationIntegerTests {
 
     static final boolean DEBUG = false;
 
-    static final ConfigurableStrategy SIMPLE_FIFO_STRATEGY = IntegerStrategy.DEFAULT.withPriorityDefinitions(NodeKey.FIFO_SEQUENCE)
-            .withParallelism(Parallelism.ONE);
+    static final ConfigurableStrategy SIMPLE_FIFO_STRATEGY = IntegerStrategy.DEFAULT.withPriorityDefinitions(NodeKey.FIFO_SEQUENCE);
 
 }